// src/redux/watchlist/watchlist.types.ts
import { CoinData } from '../../types/types';

export enum WatchlistActionTypes {
  FETCH_WATCHLIST_START = 'FETCH_WATCHLIST_START',
  FETCH_WATCHLIST_SUCCESS = 'FETCH_WATCHLIST_SUCCESS',
  FETCH_WATCHLIST_FAILURE = 'FETCH_WATCHLIST_FAILURE',

  FETCH_WATCHLIST_COINS_START = 'FETCH_WATCHLIST_COINS_START',
  FETCH_WATCHLIST_COINS_SUCCESS = 'FETCH_WATCHLIST_COINS_SUCCESS',
  FETCH_WATCHLIST_COINS_FAILURE = 'FETCH_WATCHLIST_COINS_FAILURE',

  ADD_WATCHLIST_START = 'ADD_WATCHLIST_START',
  ADD_WATCHLIST_SUCCESS = 'ADD_WATCHLIST_SUCCESS',
  ADD_WATCHLIST_FAILURE = 'ADD_WATCHLIST_FAILURE',

  REMOVE_WATCHLIST_START = 'REMOVE_WATCHLIST_START',
  REMOVE_WATCHLIST_SUCCESS = 'REMOVE_WATCHLIST_SUCCESS',
  REMOVE_WATCHLIST_FAILURE = 'REMOVE_WATCHLIST_FAILURE',
}

export interface WatchlistState {
  watched_coins: string[];
  coins: CoinData[];
  loading: boolean;
  error?: string;
}

export default WatchlistActionTypes;